Only generate a single transaction with multiple postings
It used to generate on transaction per lot (with exactly 2 postings). Now it generates a single transaction with as many postings as required to make the transfer.

This is closer conceptually to what is happening, it is a single transaction. It also makes it easier to see what is happening in bean-doctor.